Skip rp2040 decoupling test and add benchmark
